migrate `bin` generation logic to public api
