+++ pkg/generate/bin/options_test.go
@@ -1,0 +1,36 @@
+// This file is part of CycloneDX GoMod
+//
+// Licensed under the Apache License, Version 2.0 (the “License”);
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an “AS IS” BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (c) OWASP Foundation. All Rights Reserved.
+
+package bin
+
+import "testing"
+
+func TestWithLicenseResolution(t *testing.T) {
+	// TODO
+}
+
+func TestWithLogger(t *testing.T) {
+	// TODO
+}
+
+func TestWithStdlib(t *testing.T) {
+	// TODO
+}
+
+func TestWithVersionOverride(t *testing.T) {
+	// TODO
+}
